Fix cmake clang warnings

+++ ext/test/w3c_tracecontext_test/main.cc
@@ -71,7 +71,7 @@ public:
     {
       return false;
     }
-    for (int i = 0; i < str1.length(); i++)
+    for (size_t i = 0; i < str1.length(); i++)
     {
       if (tolower(str1[i]) != tolower(str2[i]))
       {
